EventCountdown: extract countdown fn to composable
