+++ src/components/EventCountdown.vue
@@ -21,14 +21,16 @@
  * <event-countdown :releaseDate="targetDate" />
  */
 
+// libraries
 import { setCssVar, date } from 'quasar';
-import { defineComponent, ref, watchEffect, onBeforeUnmount } from 'vue';
+import { defineComponent } from 'vue';
 // import { useI18n } from 'vue-i18n'
 
-// types
-import { Countdown, ConfigGlobal } from 'components/types';
+// composables
+import { useCountdown } from 'src/composables/useCountdown';
 
-const { formatDate } = date;
+// types
+import { ConfigGlobal } from 'components/types';
 
 const rideToWorkByBikeConfig: ConfigGlobal = JSON.parse(
   process.env.RIDE_TO_WORK_BY_BIKE_CONFIG
@@ -44,68 +46,16 @@ export default defineComponent({
     },
   },
   setup(props) {
-    const countdown = ref<Countdown>({
-      days: 0,
-      hours: 0,
-      minutes: 0,
-      seconds: 0,
-    });
-
     // currently not working see https://github.com/intlify/vue-i18n-next/issues/1193
     // const { locale } = useI18n({ useScope: 'global' })
     let formatString = 'D. M.';
     // if (locale.value === 'en') {
     //   formatString = 'D MMM';
     // }
+    const { formatDate } = date;
     const formattedDate = formatDate(new Date(props.releaseDate), formatString);
 
-    let countdownInterval: ReturnType<typeof setInterval> | null = null;
-
-    const startCountdown = () => {
-      const targetDate = new Date(props.releaseDate).getTime();
-
-      countdownInterval = setInterval(() => {
-        const now = new Date().getTime();
-        const timeDifference = getTimeDifference(now, targetDate);
-
-        computeCountdownInterval(timeDifference);
-      }, 1000);
-    };
-
-    function getTimeDifference(now: number, date: number): number {
-      return date - now;
-    }
-
-    function computeCountdownInterval(timeDifference: number) {
-      if (timeDifference > 0) {
-        setCountdownValues(timeDifference);
-      } else {
-        if (countdownInterval) {
-          clearInterval(countdownInterval);
-        }
-      }
-    }
-
-    function setCountdownValues(timeDifference: number): void {
-      countdown.value = {
-        days: Math.floor(timeDifference / (1000 * 60 * 60 * 24)),
-        hours: Math.floor(
-          (timeDifference % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)
-        ),
-        minutes: Math.floor((timeDifference % (1000 * 60 * 60)) / (1000 * 60)),
-        seconds: Math.floor((timeDifference % (1000 * 60)) / 1000),
-      };
-    }
-
-    watchEffect(() => {
-      startCountdown();
-    });
-
-    onBeforeUnmount(() => {
-      if (countdownInterval) {
-        clearInterval(countdownInterval);
-      }
-    });
+    const { countdown } = useCountdown(props.releaseDate);
 
     return {
       countdown,
@@ -116,12 +66,7 @@ export default defineComponent({
 </script>
 
 <template>
-  <q-card
-    square
-    flat
-    class="row items-center justify-evenly q-py-xl bg-info"
-    data-cy="card"
-  >
+  <q-card square flat class="row items-center justify-evenly q-py-xl bg-info" data-cy="card">
     <div class="text-center">
       <!-- Title -->
       <div class="text-weight-bold q-px-lg" data-cy="title">
